Allows outputting multiple CNI spec versions
This commit should allow the index2ip plugin to output any number of CNI
spec formats from 0.2.0 to 0.4.0. The main motivation for this change is
making it easier to migrate from one CNI spec version to another. This
plugin is deployed through changes to epoxy-images, but the CNI configs
are stored as ConfigMaps in the cluster (deployed through k8s-support).
Without the ability for this plugin to support multiple output formats
we would have to very closely time the rollouts of updated epoxy-images
with the rollout of k8s-support. With this change, the plugin can still
out put CNI spec 0.2.0, if the input config specifies that version. This
will allow us to deploy this updated plugin, and then later on make the
required changes to the k8s configs to start specifying a 0.3.1 output.